# Further reduce the number of copies for client-server array transfers