Eliminate a copy for `ak.array()` by ensuring that copy elision fires
when recv'ing the binary payload.

Here's the performance impact for 16-node-xc:

| config | to_ndarray | ak.array   |
| ------ | ---------: | ---------: |
| before |  715 MiB/s |  705 MiB/s |
| after  |  715 MiB/s |  750 MiB/s |

Part of #794